profile view + tables diagram

+++ src/controllers/usersController.js
@@ -21,6 +21,10 @@ const controlador = {
     res.render("registration");
   },
 
+  profile: (req, res) => {
+    res.render('profile');
+  },
+
   registration: (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -30,25 +34,35 @@ const controlador = {
         oldData: req.body,
       });
     } else {
-      let greatestId = 0;
-      users.map((user) => {
-        user.id > greatestId ? (greatestId = user.id) : '';
-      });
-      let newUser = {
-        ...req.body,
-        password: bcrypt.hashSync(req.body.password, 10),
-        image: req.file.filename,
-        id: greatestId + 1,
-      };
-      users.push(newUser);
-      fs.writeFileSync(usersFilePath, JSON.stringify(users, null, 4));
-      console.log("********* CREATION SUCCESSFUL **************");
-      return res.redirect("/");
+      if (req.body.confirmPassword == req.body.password) {
+
+        let greatestId = 0;
+        users.map((user) => {
+          user.id > greatestId ? (greatestId = user.id) : '';
+        });
+        let newUser = {
+          names: req.body.names,
+          email: req.body.email,
+          address: req.body.address,
+          password: bcrypt.hashSync(req.body.password, 10),
+          image: req.file.filename,
+          id: greatestId + 1,
+        };
+        users.push(newUser);
+        fs.writeFileSync(usersFilePath, JSON.stringify(users, null, 4));
+        console.log("********* CREATION SUCCESSFUL **************");
+        return res.render('profile', { user: newUser });
+      } else {
+        return res.render('registration', {
+          errors: 'Las contraseñas no coinciden.',
+          oldData: req.body,
+        })
+      } 
     }
   },
 
   login: (req, res) => {
-    const errors =validationResult(req);
+    const errors = validationResult(req);
     if (!errors.isEmpty()) {
       console.log(errors);
       return res.render('login', {
